Add FitResult.
FitResult contains the result of the fit: the parameters, errors, chi2
and the flag that tells if the fit went well or not. Although the
FitResults could be induced from FitPar **when the fit goes well** it
is interesting to handle the case of fit failure. In this case one
wants to ignore FitPar, but FitResults can still be returned (e.g,
taking the mean and std from the seed). It is also convenient to stick
chi2 and the valid flag here.