Remove field seed from FitCollection.
The seed is not a result of the fit, and in any case is not in general
a GaussianPar (it will be different for exponential or other fits). It
was included in FitCollection as a fallback for the cases of fit
failure, but the inclusion of FitResults allow to take this case into
account in a more general manner.