Add a test checking pulls on energy fits.
This is an important test that computes the pulls between true and
fitted values for both the fit and the average/rms. The values of
sigmas and means are varied as well as the number of experiments. The
mean of the fits are unbiased and the pulls are correct, while the
means of the stds are slighly biased (i 1 sigma), and this is taken
into account in the test.