32 Fix run_number type to be integer.
[author: ausonandres]
This PR changes the run_number type in map building config. file to be
an integer, in order to avoid an error in Esmeralda when comparing
that value to 0, to apply temporal corrections in data

[reviewer: bpalmeiro]
This is a temporal issue as Esmeralda does not accept strings for run
number. This should be undone when the city changes that. So far, it
is necessary. Good job.